+++ .storybook/main.js
@@ -34,7 +34,7 @@ module.exports = {
           loader: "css-loader",
           options: {
             importLoaders: 1,
-            modules: true,
+            modules: { auto: true },
           },
         },
         "postcss-loader",