+++ packages/core/src/theme/DefaultTheme.ts
@@ -64,8 +64,8 @@ export const defaultTheme: Theme = {
     tiny: "9px",
   },
   shadows: {
-    modal: "rgba(0, 0, 0, 0.2) 0px 0px 10px 4px;",
-    popover: "rgba(0, 0, 0, 0.2) 0px 0px 6px 4px;",
-    box: "rgba(0, 0, 0, 0.15) 0 2px 4px 0;",
+    modal: "var(--swui-shadow-modal)",
+    popover: "var(--swui-shadow-popover)",
+    box: "var(--swui-shadow-box)",
   },
 };