+++ packages/grid/src/features/standard-table/components/TextCell.tsx
@@ -1,14 +1,19 @@
 import { Indent, Text } from "@stenajs-webui/core";
 import * as React from "react";
+import styles from "./TextCell.module.css";
 
 interface Props {
   label?: string;
 }
 
-export const TextCell: React.FC<Props> = React.memo(function ({ label }) {
+export const TextCell: React.FC<Props> = React.memo(function TextCell({
+  label,
+}) {
   return (
-    <Indent>
-      <Text>{label}</Text>
+    <Indent overflow={"hidden"}>
+      <Text className={styles.textCell} title={label}>
+        {label}
+      </Text>
     </Indent>
   );
 });